requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
kubernetes==28.1.0
pyyaml==6.0.1
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.1